backend/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
pymongo==4.6.2
python-dotenv==1.0.1
PyJWT==2.8.0
bcrypt==4.1.2
python-jose==3.3.0